requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
black==23.3.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cx-Oracle==8.3.0
decorator==5.1.1
Django==4.2.5
django-cors-headers==3.14.0
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
executing==1.2.0
fpdf==1.7.2
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
ipdb==0.13.13
ipython==8.13.1
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
mysqlclient==2.2.0
numpy==1.26.0
openapi-codec==1.3.2
packaging==23.1
pandas==2.1.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
PyJWT==2.8.0
python-barcode==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-environ==0.4.54
python-magic==0.4.27
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sentry-sdk==1.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
XlsxWriter==3.1.6
